order, product -> auth 미들웨어 추가 및 삭제,수정 권한 로직 추가

+++ routes/order.js
@@ -8,10 +8,11 @@ router.use(express.urlencoded({ extended: true }));
 
 // 주문하기
 router.post("/order", auth, async (req, res) => {
+    const { user } = res.locals;
     try {
         for (let order of req.body) {
-            const { orderer, productId, productName, orderQuantity } = order;
-            await Order.create({ orderer, productId, productName, orderQuantity });
+            const { productId, productName, orderQuantity } = order;
+            await Order.create({ orderer: user.userId, productId, productName, orderQuantity });
         }
         res.status(200).send({ message: "주문이 접수되었습니다." });
     } catch (error) {
@@ -22,12 +23,17 @@ router.post("/order", auth, async (req, res) => {
 // 주문 부분 취소
 router.delete("/order/:orderId", auth, async (req, res) => {
     const { orderId } = req.params;
+    const { user } = res.locals;
     const order = await Order.findOne({ _id: orderId });
     try {
-        if (order.orderState === "배송완료") 
-            return res.status(200).send({ message: "이미 배송이 완료된 상품입니다." });
-        await Order.deleteOne({ _id: orderId });
-        res.status(200).send({ message: "해당 주문을 취소했습니다." });
+        if (user.userId === order.orderer) {
+            if (order.orderState === "배송완료")
+                return res.status(200).send({ message: "이미 배송이 완료된 상품입니다." });
+            await Order.deleteOne({ _id: orderId });
+            res.status(200).send({ message: "해당 주문을 취소했습니다." });
+        } else {
+            res.status(400).send({ message: "삭제 권한이 없습니다." });
+        }
     } catch (error) {
         return res.status(400).send({ message: "올바르지 않은 접근입니다." });
     }
